pyproject.toml
[project]
name = "django-jwt-header-auth"
version = "0.1.0"
description = "Django authentication backend and middleware for JWT header authentication"
requires-python = ">=3.14,<3.15"
dependencies = [
    "django>=5.2",
    "pyjwt>=2.10.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jwt_auth"]
